requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
scikit-learn==1.5.2
numpy==2.1.1
pyyaml==6.0.2
scipy==1.14.1
joblib==1.4.2
pytest==8.3.3
